unload `abstract-command`
